ToolChain update. GCC 5.2 with new libs and fixes, made for 3.4.110 and UP.

+++ android-toolchain/arm-LG-linux-gnueabi/sysroot/usr/include/drm/sis_drm.h
@@ -51,17 +51,17 @@
 
 typedef struct {
 	int context;
-	unsigned long offset;
-	unsigned long size;
+	unsigned int offset;
+	unsigned int size;
 	unsigned long free;
 } drm_sis_mem_t;
 
 typedef struct {
-	unsigned long offset, size;
+	unsigned int offset, size;
 } drm_sis_agp_t;
 
 typedef struct {
-	unsigned long offset, size;
+	unsigned int offset, size;
 } drm_sis_fb_t;
 
 struct sis_file_private {
